Fixed doctrine persistence namespace rename
