+++ src/Bridge/Doctrine/Common/ObjectWriter.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Yokai\Batch\Bridge\Doctrine\Common;
 
-use Doctrine\Common\Persistence\ManagerRegistry;
-use Doctrine\Common\Persistence\ObjectManager;
+use Doctrine\Persistence\ManagerRegistry;
+use Doctrine\Persistence\ObjectManager;
 use Yokai\Batch\Job\Item\ItemWriterInterface;
 
 final class ObjectWriter implements ItemWriterInterface
