+++ src/Bridge/Doctrine/Orm/EntityReader.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Yokai\Batch\Bridge\Doctrine\Orm;
 
-use Doctrine\Common\Persistence\ManagerRegistry;
+use Doctrine\Persistence\ManagerRegistry;
 use Doctrine\ORM\EntityManagerInterface;
 use Yokai\Batch\Job\Item\ItemReaderInterface;
