+++ tests/unit/Bridge/Doctrine/Common/ObjectWriterTest.php
@@ -2,8 +2,8 @@
 
 namespace Yokai\Batch\Tests\Unit\Bridge\Doctrine\Common;
 
-use Doctrine\Common\Persistence\ManagerRegistry;
-use Doctrine\Common\Persistence\ObjectManager;
+use Doctrine\Persistence\ManagerRegistry;
+use Doctrine\Persistence\ObjectManager;
 use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
 use Prophecy\Prophecy\ObjectProphecy;
